+++ neuropacks/utils/binning.py
@@ -1,7 +1,8 @@
 import numpy as np
 
-from scipy.interpolate import interp1d
 from scipy.ndimage import gaussian_filter1d
+
+from neuropacks.utils.signal import box_cox
 
 
 def create_bins(t_start, t_end, bin_width, bin_type='time'):
@@ -122,14 +123,3 @@ def build_filter(gaussian=None):
         return x_traj
 
     return filter
-
-
-def box_cox(x, power_param):
-    ''' one-parameter Box-Cox transformation '''
-    return (np.power(x, power_param) - 1) / power_param
-
-
-def downsample_by_interp(x, t, t_samp):
-    interpolator = interp1d(t, x)
-    x_samp = interpolator(t_samp)
-    return x_samp
