+++ neuropacks/utils/io.py
@@ -1,0 +1,5 @@
+import pickle
+
+
+def load_pickle(filename):
+    return pickle.load(open(filename, 'rb'))
